LPD-20229 Move blocks folder
