+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Solutions/components/Blocks/ImagesGrid.tsx
@@ -3,18 +3,20 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {DropzoneUpload} from '../../../../../../../components/DropzoneUpload/DropzoneUpload';
-import {ACCEPT_FILE_TYPES} from '../../../../../../StorefrontPage/CustomizeAppStorefrontPage';
-import TextBlock from './TextBlock';
+import {DropzoneUpload} from '../../../../../../components/DropzoneUpload/DropzoneUpload';
+import Form from '../../../../../../components/MarketplaceForm';
+import {ACCEPT_FILE_TYPES} from '../../../../../StorefrontPage/CustomizeAppStorefrontPage';
 
-const TextAndImages = () => {
-	const handleUpload = (files: File[]) => <div>{files[0].name}</div>;
+const ImagesGrid = () => {
+	const handleUpload = (_files: File[]) => null;
 
 	return (
 		<>
-			<TextBlock />
+			<Form.Label className="mt-1 p-4" htmlFor="description">
+				Add up to 6 images
+			</Form.Label>
 
-			<div className="d-flex p-4">
+			<div className="d-flex mb-4 px-4">
 				<DropzoneUpload
 					acceptFileTypes={ACCEPT_FILE_TYPES}
 					buttonText="Select a file"
@@ -30,4 +32,4 @@ const TextAndImages = () => {
 	);
 };
 
-export default TextAndImages;
+export default ImagesGrid;
