+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Solutions/components/Blocks/SingleImage.tsx
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {DropzoneUpload} from '../../../../../../../components/DropzoneUpload/DropzoneUpload';
-import {ACCEPT_FILE_TYPES} from '../../../../../../StorefrontPage/CustomizeAppStorefrontPage';
+import {DropzoneUpload} from '../../../../../../components/DropzoneUpload/DropzoneUpload';
+import {ACCEPT_FILE_TYPES} from '../../../../../StorefrontPage/CustomizeAppStorefrontPage';
+import {MAX_SIZE_5MBS} from '../../constants';
 
 const SingleImage = () => {
-	const handleUpload = (files: File[]) => <div>{files[0].name}</div>;
+	const handleUpload = (_files: File[]) => null;
 
 	return (
 		<div className="d-flex p-4">
@@ -16,7 +17,7 @@ const SingleImage = () => {
 				buttonText="Select a file"
 				description="Only gif, jpg, png are allowed. Max file size is 5MB "
 				maxFiles={5}
-				maxSize={5000000}
+				maxSize={MAX_SIZE_5MBS}
 				multiple={true}
 				onHandleUpload={handleUpload}
 				title="Drag and drop to upload or"
